Move configuration objects into a separated package for better project organizing.
