+++ src/test/java/org/masukomi/aspirin/core/DefunctModuleTest.java
@@ -13,6 +13,7 @@ import javax.mail.internet.MimeMessage;
 import javax.management.MBeanServer;
 import javax.management.ObjectName;
 
+import org.masukomi.aspirin.core.config.Configuration;
 import org.masukomi.aspirin.core.store.mail.FileMailStore;
 
 import junit.framework.TestCase;
