+++ www/configuration.html
@@ -25,7 +25,7 @@ Configuring Aspirin</span><br>
 		<li>aspirinMaxAttempts - the maximum number of times it will try and send a message if it has problems.</li>
 	</ul>
 <p><b>Configuring via setter methods</b><br />
-	See the Javadocs in <tt>org.masukomi.aspirin.core.Configuration</tt>for details on this method.  </p>
+	See the Javadocs in <tt>org.masukomi.aspirin.core.config.Configuration</tt>for details on this method.  </p>
 <p><b>Configuring via doing nothing</b><br />
 	See above for the default settings. </p>
 <p><b>Configuring Aspirin's Logging</span><br>
